chatbot/requirements.txt
huggingface-hub==0.2.1 
ipykernel==6.6.0 
ipython==7.30.1
numpy==1.21.4 
pandas==1.3.5 
pytorch-lightning==1.5.6
tokenizers==0.10.3
torchaudio==0.10.1
torchmetrics==0.6.2
torchvision==0.11.2
tqdm==4.62.3
transformers==4.14.1
urllib3==1.26.7